refactor: move `splitSegments` to another file
