+++ packages/backend/src/core/MfmService.ts
@@ -14,6 +14,7 @@ import { intersperse } from '@/misc/prelude/array.js';
 import { normalizeForSearch } from '@/misc/normalize-for-search.js';
 import type { IMentionedRemoteUsers } from '@/models/Note.js';
 import { bindThis } from '@/decorators.js';
+import { splitSegments } from '@/misc/split-segments.js';
 import type { DefaultTreeAdapterMap } from 'parse5';
 import type * as mfm from 'mfm-js';
 
@@ -248,50 +249,6 @@ export class MfmService {
 				// otherwise, we don't need to escape
 				return text;
 			}
-		}
-
-		/**
-		 * textをregexesで分割するが、分割するときに regexes にマッチした部分も含める。
-		 */
-		function splitSegments(text: string, regexes: RegExp[]): [regexIdx: number, text: string][] {
-			const result: [regexIdx: number, text: string][] = [];
-
-			let rest = text;
-			for (;;) {
-				let matchRegex: [number, RegExpExecArray] | null = null;
-
-				for (let i = 0; i < regexes.length; i++) {
-					const regex = regexes[i];
-					regex.lastIndex = 0;
-					const matchCurrent = regex.exec(rest);
-					if (matchCurrent) {
-						if (matchRegex != null) {
-							if (matchCurrent.index < matchRegex[1].index) {
-								matchRegex = [i, matchCurrent];
-							}
-						} else {
-							matchRegex = [i, matchCurrent];
-						}
-					}
-				}
-
-				if (matchRegex != null) {
-					const [i, match] = matchRegex;
-
-					const head = rest.slice(0, match.index);
-					const segment = match[0];
-					const tail = rest.slice(match.index + segment.length);
-
-					result.push([-1, head]);
-					result.push([i, segment]);
-					rest = tail;
-				} else {
-					result.push([-1, rest]);
-					break;
-				}
-			}
-
-			return result;
 		}
 
 		const emojiCodeRegex = /(?<![a-z0-9]):[a-z0-9_]+:(?![a-z0-9])/i;
